move CsvFileSplitter, to stop sonarqube complaining about duplicate code
