+++ metrics_utility/base/__init__.py
@@ -1,7 +1,6 @@
 from .collection_csv import CollectionCSV
 from .collection_json import CollectionJSON
 from .collector import Collector
-from .csv_file_splitter import CsvFileSplitter
 from .decorators import register
 from .package import Package
 
@@ -9,7 +8,6 @@ from .package import Package
 __all__ = [
     'Collector',
     'Package',
-    'CsvFileSplitter',
     'CollectionCSV',
     'CollectionJSON',
     'register',
